+++ lib/hexapdf/layout/container_box.rb
@@ -163,15 +163,13 @@ module HexaPDF
         @box_fitter = BoxFitter.new([my_frame])
         children.each {|box| @box_fitter.fit(box) }
 
-        if @box_fitter.success?
+        if @box_fitter.success? || (!@box_fitter.fit_results.empty? && @splitable)
           update_content_width do
             result = @box_fitter.fit_results.max_by {|r| r.mask.x + r.mask.width }
             children.empty? ? 0 : result.mask.x + result.mask.width - my_frame.left
           end
           update_content_height { @box_fitter.content_heights.max }
-          fit_result.success!
-        elsif !@box_fitter.fit_results.empty? && @splitable
-          fit_result.overflow!
+          @box_fitter.success? ? fit_result.success! : fit_result.overflow!
         end
       end